CORE-7071: Move to a new module `rbac-schema`
