+++ data/rbac-schema/README.md
@@ -1,0 +1,4 @@
+# Role Based Access Control (RBAC) Schema
+
+This project contains useful constants which are used in RBAC sub-system for validating HTTP users permissions
+entitlements.
