+++ data/rbac-schema/src/main/java/net/corda/rbac/schema/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.rbac.schema;
+
+import org.osgi.annotation.bundle.Export;
