+++ settings.gradle
@@ -97,6 +97,7 @@ include 'data:config-schema'
 include 'data:db-schema'
 include 'data:membership-schema'
 include 'data:topic-schema'
+include 'data:rbac-schema'
 include 'ledger:ledger-common'
 include 'ledger:ledger-consensual'
 include 'ledger:ledger-utxo'
